requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Dance>=7.1.0
Flask-Mail>=0.10.0
python-dotenv>=1.0.0
Werkzeug==2.3.7
gunicorn>=23.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib>=1.5.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
bleach==6.0.0
cryptography==41.0.4
PyOpenSSL>=25.1.0
ssl==1.16
PyJWT>=2.10.1
oauthlib>=3.3.1
urllib3==2.0.4
whois==0.9.27
python-whois>=0.9.5
dnspython==2.4.2
tldextract==3.4.4
validators==0.22.0
email-validator==2.0.0
SQLAlchemy==2.0.21
Alembic==1.12.0
psycopg2-binary>=2.9.10
greenlet==2.0.2
Jinja2==3.1.2
WTForms==3.0.1
MarkupSafe==2.1.3
Mako==1.2.4
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
pytz==2023.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.7.1
future==0.18.3
charset-normalizer==3.2.0
certifi==2023.7.22
idna==3.4
re2==1.1
reportlab==4.0.4
Pillow==10.0.0







          
## Library Names with Single Line Descriptions

Here's a comprehensive list of all libraries used in your Automate Phishing Detection System with brief descriptions:

### **Web Framework & Core**
- **Flask** - Lightweight Python web framework for building web applications
- **Flask-SQLAlchemy** - Flask extension that adds SQLAlchemy ORM support
- **Flask-Login** - User session management and authentication for Flask
- **Flask-WTF** - Integration of Flask and WTForms for form handling
- **Flask-Migrate** - Database migration support for Flask applications
- **Flask-Dance** - OAuth consumer extension for Flask applications
- **Flask-Mail** - Email sending capabilities for Flask applications
- **Werkzeug** - WSGI utility library that powers Flask
- **Gunicorn** - Python WSGI HTTP Server for UNIX systems

### **Machine Learning & Data Analysis**
- **scikit-learn** - Machine learning library with classification and regression algorithms
- **pandas** - Data manipulation and analysis library with DataFrame structures
- **numpy** - Fundamental package for scientific computing with arrays
- **joblib** - Lightweight pipelining and parallel computing library

### **Web Scraping & HTML Processing**
- **requests** - HTTP library for making web requests and API calls
- **beautifulsoup4** - HTML and XML parsing library for web scraping
- **lxml** - Fast XML and HTML parser with XPath support
- **html5lib** - Pure-Python HTML5 parser based on WHATWG specification
- **bleach** - HTML sanitization library to prevent XSS attacks

### **Security & Cryptography**
- **cryptography** - Cryptographic recipes and primitives for Python
- **PyOpenSSL** - Python wrapper around OpenSSL library for SSL/TLS
- **ssl** - SSL/TLS wrapper for socket objects (built-in module)
- **PyJWT** - JSON Web Token implementation for Python
- **oauthlib** - Generic OAuth request-signing logic implementation

### **Network & Domain Analysis**
- **urllib3** - HTTP client library with connection pooling and SSL/TLS verification
- **whois** - Python library for retrieving WHOIS information about domains
- **python-whois** - Alternative WHOIS client library for domain lookups
- **dnspython** - DNS toolkit for Python with query and zone file support
- **tldextract** - Library to extract domain components from URLs
- **validators** - Data validation library for URLs, emails, and other formats
- **email-validator** - Robust email address syntax and deliverability validation

### **Database**
- **SQLAlchemy** - SQL toolkit and Object-Relational Mapping library
- **Alembic** - Database migration tool for SQLAlchemy
- **psycopg2-binary** - PostgreSQL database adapter for Python
- **greenlet** - Lightweight in-process concurrent programming library

### **Template & Form Processing**
- **Jinja2** - Modern templating engine for Python applications
- **WTForms** - Flexible forms validation and rendering library
- **MarkupSafe** - String handling library that implements automatic string escaping
- **Mako** - Fast template library with Python-like syntax

### **Utilities & Support**
- **click** - Command line interface creation toolkit
- **itsdangerous** - Data signing and secure cookie implementation
- **blinker** - Fast Python in-process signal/event dispatching system
- **pytz** - World timezone definitions and utilities
- **python-dateutil** - Extensions to the standard Python datetime module
- **six** - Python 2 and 3 compatibility utilities
- **typing_extensions** - Backported type hints for older Python versions
- **future** - Clean single-source support for Python 3 and 2

### **Text & Character Processing**
- **charset-normalizer** - Character encoding detection library
- **certifi** - Collection of Root Certificates for validating SSL certificates
- **idna** - Internationalized Domain Names in Applications implementation
- **re2** - Fast regular expression library with linear time complexity

### **Document & Image Processing**
- **reportlab** - PDF generation library for creating documents and reports
- **Pillow** - Python Imaging Library fork for image processing


        